+++ RAnalogs/RAnEn/src/RcppChecks.cpp
@@ -27,3 +27,25 @@ bool checkOpenMP() {
 #endif
 }
 
+//' @export
+// [[Rcpp::export]]
+void setNumThreads(int threads) {
+#ifndef _OPENMP
+  omp_set_num_threads(threads);
+#else
+  Rcpp::Rcout << "Nothing happened because OpenMP is not supported." << std::endl;
+#endif
+  return;
+}
+
+//' @export
+// [[Rcpp::export]]
+int getNumThreads() {
+#ifndef _OPENMP
+  return omp_get_num_procs();
+#else
+  Rcpp::Rcout << "Only using a single thread because OpenMP is not supported." << std::endl;
+  return 1;
+#endif
+}
+