+++ NEWS.md
@@ -3,6 +3,8 @@
 - Removed profiling current memory
 - Added sorting capability to `RAnEn::formatObservations`
 - Added argument to read a subset of variables to `RAnEn::readNc`
+- Bug fixed in `RAnEn::formatObservations` when X and Y column names are used
+- Bug fixed in `RAnEn::writeNetCDF` when `StationNames` is empty
 
 # PAnEn 4.0.6