+++ NEWS.md
@@ -1,3 +1,7 @@
+# PAnEn 4.0.8
+
+- Added thread control
+
 # PAnEn 4.0.7
 
 - Removed profiling current memory